+++ packages/adp-tooling/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @sap-ux/adp-tooling
 
+## 0.12.13
+
+### Patch Changes
+
+-   4c06318: Add Adaptation Project's Add Annotations To OData generator prompting
+
+## 0.12.12
+
+### Patch Changes
+
+-   Updated dependencies [108336f]
+    -   @sap-ux/inquirer-common@0.4.1
+
 ## 0.12.11
 
 ### Patch Changes